put failover specific exceptions in own file
(cherry picked from commit 2a183f68b94caa49da3fc9fa5874763f1a0e2989)
